requirements.txt
python-dotenv==1.0.1
Django==5.1.6
uuid==1.30
pillow==10.4.0
psycopg-binary==3.2.5
django-extensions==3.2.3
django_rest_framework
django-silk==5.3.2